Repository interface (#99)
Currently, we have a repository interface for every document. But the underlying code is almost the same for each document repository.

Moving to more abstracted implementation will not only leads to less repetitive code but easier to migrate and manage multiple repository implementations.

This PR defines an abstract and implementation for `purchaseorderdocument`
Once approved, we move the remaining document repositories under the same abstraction.